dynamic report based on completed steps
The "done" statement won't work with MultiQC, potentially because of the wrapper, I think because it sets a working directory and doesn't like the fact that I try to touch the file somewhere else. Workaround: use the multiqc html report as done file. Now, report is generated upon completion of the workflow, no matter which step has completed. Note: for excluded steps, there's no mention whatsoever in the report