check if log files of cutadapt, pandaseq and vsearch are properly formatted
related to the version of each package. If the user launches the workflow with 'snakemake' only instead of 'snakemake --use-conda', it can lead to the use of packages available from the main env, most likely the wrong versions, which will lead to cryptic errors (e.g. related to formatting of the log files)