re-design of how the mutant layout is generated
Mutant layout is obtained by first generating a dataframe containing all expected mutants. This step is optional and for barcoded designs, we ask the user to provide the dataframe of barcode-variant associations. Regardless of the file provenance (user-provided or generated by the workflow), the dataframe is then annotated, among other things to identify mutations. This should be pretty flexible but I expect it to be greedy in computational resources for complex designs (multiple loci with sequences with many mutations). More tests coming soon.