patch bd049da
Important note: the concept behing collect_graphs() doesn't really work because snakemake auto-collects **all files with a report() directive**. The targets for the --report command are only for evaluation (DAG) so that it auto-runs the missing jobs (which in our case is futile). There's currently no way to filter out some files from the report if they were created with a report() directive. We could simply not generate the files (by expanding the wildcard differently), but I think it's safer to generate them and include a blank graph if the samples weren't marked for reporting. Maybe a future report plugin will solve this limitation.